+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add Request Archive form - #832
 - add Request Solve form - #830
 - add Request Add form - #833
+- add fragment #card upon offer desactivation & re-activation - #978
 
 ### Changed
 - update all Javascript Dependencies
@@ -34,6 +35,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix un-peristed contact firstname & lasname upon offer creation - #944
 - fix un-peristed contact firstname & lasname upon request creation
 - fix wrong persistence of e-mail & phone number upon request creation
+- fix counter of offers by types - #981
 
 ### Added
 - add Docker credentials for Codeship to prevent Pull Rate Limit